Refactor to make better room for process metrics
I introduced the Collector abstraction, as is already suggested by the
Prometheus documentation and technically another abstraction level just
before serialization. I don't know if I like this because it's very memcpy
intensive now, but it does make room for process metrics fairly nicely.
It's not quite done (see TODO at the bottom of the Registry), but it's
too late to finish it today.